enable FP16 inference

+++ ldm/modules/diffusionmodules/util.py
@@ -216,7 +216,8 @@ class SiLU(nn.Module):
 
 class GroupNorm32(nn.GroupNorm):
     def forward(self, x):
-        return super().forward(x.float()).type(x.dtype)
+        # return super().forward(x.float()).type(x.dtype)
+        return super().forward(x).type(x.dtype)
 
 def conv_nd(dims, *args, **kwargs):
     """
